Add overrides to machine specializations
Currently, there are four specializations that only accept that external mods cannot interact with or add input items to. The most important of these are:
1. The Induction Smelter's *Pyro-Concentrator* specialization.
2. The Pulverizer's *Tectonic Initiator* specialization.
3. The Redstone Furnace's *Trivection Chamber* and *Flux Anodizers* specialization.

This pull request adds methods to add overrides to the helper methods that these specializations rely on to let external code mark items as valid inputs for these specializations.
Originally, I was going to add a seperate recipe map for each of these specializations; however, after some thought I concluded that this would be an inappropriate course of action considering the popularity and importance that this mod has, as this would require significant changes that would likely create incompatibilities with pre-existing mods that rely on Thermal Expansion.

Only four existing codepoints were touched in this request: `FurnaceManager.isFood(stack)`, `FurnaceManager.isOre(stack)`, `PulverizerManager.isOre(stack)` and `SmelterManager.isOre(stack)`. The only other changes are methods being added to interface with the override maps. The mods has been built and tested to ensure that the existing behaviours remain unmodified.

If this pull request is accepted, I will likely immediately create a second pull request for ModTweaker to add this functionality.